Waiting time validator fixes (#723)
* Handling of missing ScheduledStopPoint for Interchange

Ensures InterchangeWaitingTimeValidator handles the cases where the ScheduledStopPointRef referred to by a FromPointRef or ToPointRef is not referred to by any of the StopPointInJourneyPatterns for that Service Journey/Journey Pattern

* Gets operating day dates for common file via the service calendar

Co-authored-by: eibakke <eivind.morris.bakke@entur.org>

* Removes follow-up errors in ValidationReport

Removes a case of follow-up errors produced in ValidationReport by InterchangeWaitingTimeValidator due to ScheduledStopPoint references missing in Service Journeys. These errors are already handled by StopPointInInVehicleJourneyValidator.

Adds relevant docs to interchange validation code.

* Fixes NPEs in active dates collector

* Delegates ServiceCalendar parsing to netex-parser-java

* Fixes broken active dates collector for DSJs

* Fixes NPE due to missing departureTime in waiting time validator

* Enables interchange waiting time validation in dev

* Adds better error handling to dataset validation

---------

Co-authored-by: eibakke <eivind.morris.bakke@entur.org>