Adds handling of ServiceJourneyInterchanges referring to non-existing ServiceJourneys (#752)
* Reduces log verbosity in interchange validator

* Improves error handling in interchange validator

Ensures that datasets with interchanges referring to service journeys that do not exist, do not fail with SYSTEM_ERROR when running InterchangeWaitingTimeValidator

* Simplifies code

* Adds tests to InterchangeWaitingTimeValidatorTest

Adds tests to ensure that references to non existent service journeys from interchange do not throw an error when validating interchange waiting times

* Fixes test that didnt test anything

* Fixes ambiguous errors in StopPointsInVehicleJourneyValidator

Interchanges referring to non existing from- or toServiceJourneys no longer result in warnings from StopPointsInVehicleJourneyValidator

* Clean-up of log messages in interchange validation

* Adds InterchangeServiceJourneyReferencesExistValidator

This new validator checks that any references to ToServiceJourney and FromServiceJourney in a ServiceJourneyInterchange exist in the dataset. If they don't, a validation issue (warning) is produced in the resulting validation report.

* Fixes log statements

* Fixes issues reported by Sonar